+++ src/tests/test_cases/test.py
@@ -14,7 +14,7 @@ expect('class_', 'D101: Missing docstring in public class')
 
 class class_:
 
-    expect('meta', 'D101: Missing docstring in public class')
+    expect('meta', 'D106: Missing docstring in public nested class')
 
     class meta:
         """"""
@@ -27,6 +27,10 @@ class class_:
         pass
 
     @expect('D102: Missing docstring in public method')
+    def __new__(self=None):
+        pass
+
+    @expect('D107: Missing docstring in __init__')
     def __init__(self=None):
         pass
 
@@ -271,6 +275,10 @@ if sys.version_info[0] <= 2:
     def unicode_unmarked():
         """Юникод."""
 
+    @expect('D302: Use u""" for Unicode docstrings')
+    def first_word_has_unicode_byte():
+        """あy."""
+
 
 @expect("D400: First line should end with a period (not 'y')")
 def lwnlkjl():
@@ -375,6 +383,7 @@ def docstring_ignore_violations_of_pydocstyle_D400_and_PEP8_E501_but_catch_D401(
 def bad_decorated_function():
     """Bad (E501) but decorated"""
     pass
+
 
 expect(os.path.normcase(__file__ if __file__[-1] != 'c' else __file__[:-1]),
        'D100: Missing docstring in public module')